prep for better testing
