+++ .npmignore
@@ -21,3 +21,4 @@ docker-compose.yml
 docker-compose-toolbox.yml
 docker-compose-traefik.yml
 docker-compose-traefik-letsencrypt.yml
+.nyc_output
