+++ .nyrc.json
@@ -1,0 +1,17 @@
+{
+    "extends": "@istanbuljs/nyc-config-typescript",
+    "include": [
+        "OpenFlow/src/**/*.ts"
+    ],
+    "exclude": [
+        "node_modules/"
+    ],
+    "extension": [
+        ".ts"
+    ],
+    "reporter": [
+        "text-summary",
+        "html"
+    ],
+    "report-dir": "./coverage"
+}
