+++ OpenFlow/src/otelspec.ts
@@ -24,6 +24,7 @@ export declare class LicenseFile {
     private _ofid;
     ofid(): any;
     validate(): void;
+    shutdown(): void;
     private validateTimer;
     /**
      *  Generate license file
@@ -90,4 +91,5 @@ export declare class otel {
     startTimer(): HrTime;
     endTimer(startTime: HrTime, recorder: ValueRecorder, labels?: Labels): any;
     setdefaultlabels(): void;
+    shutdown(): Promise<void>;
 }
