+++ package.json
@@ -4,7 +4,9 @@
   "description": "Simple wrapper around NodeRed, RabbitMQ and MongoDB to support a more scaleable NodeRed implementation.\r Also the \"backend\" for [OpenRPA](https://github.com/skadefro/OpenRPA)",
   "main": "index.js",
   "scripts": {
-    "test": "ts-mocha --paths -p OpenFlow/tsconfig.json OpenFlow/**/*.spec.ts",
+    "test": "./node_modules/.bin/_mocha 'test/**/*.test.ts'",
+    "test2": "nyc ./node_modules/.bin/_mocha 'test/**/*.test.ts'",
+    "oldtest": "ts-mocha --paths -p OpenFlow/tsconfig.json OpenFlow/**/*.spec.ts",
     "build": "webpack --mode production"
   },
   "repository": {
@@ -36,6 +38,7 @@
     "@opentelemetry/exporter-collector-grpc": "0.19.0",
     "@opentelemetry/metrics": "0.19.0",
     "@opentelemetry/tracing": "0.19.0",
+    "@testdeck/mocha": "^0.1.2",
     "@types/node": "^15.12.4",
     "amqplib": "^0.8.0",
     "async-retry": "^1.3.1",
@@ -47,6 +50,7 @@
     "cookie-parser": "^1.4.5",
     "cookie-session": "^1.4.0",
     "dockerode": "^3.3.0",
+    "dotenv": "^10.0.0",
     "envfile": "^6.14.0",
     "express": "^4.17.1",
     "flash": "^1.1.0",
@@ -62,6 +66,7 @@
     "morgan": "^1.10.0",
     "multer": "^1.4.2",
     "multer-gridfs-storage": "^5.0.0",
+    "nyc": "^15.1.0",
     "oauth2-server": "^3.1.1",
     "oidc-provider": "^6.31.0",
     "passport": "^0.4.1",
@@ -76,6 +81,8 @@
     "saml20": "^0.1.14",
     "samlp": "^6.0.2",
     "shepherd.js": "^8.3.1",
+    "ts-mockito": "^2.6.1",
+    "ts-node": "^10.0.0",
     "winston": "^3.3.3",
     "ws": "^7.5.0"
   },
