+++ test/tsconfig.json.bin
@@ -1,0 +1,16 @@
+{
+    "extends": "../OpenFlow/tsconfig.json",
+    "compilerOptions": {
+        "experimentalDecorators": true,
+        "isolatedModules": false,
+        "typeRoots": [
+            "../node_modules/@types"
+        ]
+    },
+    "exclude": [
+        "../node_modules"
+    ],
+    "include": [
+        "./**/*.ts"
+    ]
+}
